fix(msvc): cast uint32_t to uint64_t in uint128_t multiply
uint32_t matches both operator*(int) and operator*(uint64_t)
equally on MSVC (narrowing vs widening). Explicit cast resolves.
Same fix in both AmmPool.cpp and PoolAMM.cpp.

+++ src/CryptoNoteCore/AmmPool.cpp
@@ -116,7 +116,7 @@ bool ammValidateDepositRatio(uint64_t amountA, uint64_t amountB,
   uint128_t delta = (expectedRatio > actualRatio) ?
     (expectedRatio - actualRatio) : (actualRatio - expectedRatio);
 
-  uint128_t maxDelta = expectedRatio * toleranceBps / FEE_DIVISOR;
+  uint128_t maxDelta = expectedRatio * (uint64_t)toleranceBps / FEE_DIVISOR;
   return delta <= maxDelta;
 }
