+++ src/SwapDaemon/PoolAMM.cpp
@@ -208,7 +208,7 @@ bool poolValidateDepositRatio(uint64_t amountA,
   uint128_t delta = (expectedRatio > actualRatio)
     ? (expectedRatio - actualRatio) : (actualRatio - expectedRatio);
 
-  uint128_t maxDelta = expectedRatio * toleranceBps / FEE_DIVISOR;
+  uint128_t maxDelta = expectedRatio * (uint64_t)toleranceBps / FEE_DIVISOR;
   return delta <= maxDelta;
 }
